Ensure lockfile is up-to-date
Fail installation if the yarn.lock file is not in sync with
packages.json. This forces any package.json changes to have a matching
yarn.lock change.  Also prevents deploying apps with dependencies that
do not matches the the truth in yarn.lock.

From yarn docs:

`yarn install --frozen-lockfile`
Don’t generate a yarn.lock lockfile and fail if an update is needed.