Scan only runtime dependencies with OWSAP dependency checker
Many projects are running with default OWASP checker settings which
means a lot of non runtime configurations are scanned, generating a lot
of noise.

I have also directly configured the checker to fail the build if a CVE
is found, rather than passing a flag to the project build script and
relying on the project to respect it.